trying to address review request

+++ addons/flannel/0.21.0/install.sh
@@ -256,9 +256,11 @@ function flannel_already_applied() {
     # We will remove the flannel pods to let it be re-created
     # in order to workaround the issue scenario described in
     # https://github.com/flannel-io/flannel/issues/1721
-    log "Deleting pods from kube-flannel namespace"
-    kubectl delete --all pods --namespace=kube-flannel
-    sleep 60
+    if [ "$KUBERNETES_UPGRADE" == "1" ]; then
+       log "Deleting pods from kube-flannel namespace"
+       kubectl delete --all pods --namespace=kube-flannel
+       sleep 60
+    fi
 
     flannel_ready_spinner
     check_network
